feat(Collapse): rewrite label from string to ReactNode
closes #2712
