+++ src/components/Collapse/CollapseIcon/CollapseIcon.tsx
@@ -3,35 +3,10 @@ import './CollapseIcon.css';
 import React, { useRef } from 'react';
 import { Transition } from 'react-transition-group';
 
-import { cnIcon, IconProps, IconPropSize } from '##/icons/Icon';
+import { cnIcon } from '##/icons/Icon';
 import { cn } from '##/utils/bem';
-import { PropsWithHTMLAttributes } from '##/utils/types/PropsWithHTMLAttributes';
 
-export const collapseIconPropDirection = [
-  'up',
-  'right',
-  'down',
-  'left',
-  'upRight',
-  'downRight',
-  'upLeft',
-  'downLeft',
-] as const;
-export type CollapseIconPropDirection =
-  typeof collapseIconPropDirection[number];
-
-export type CollapseIconProps = PropsWithHTMLAttributes<
-  {
-    size: IconPropSize;
-    icon: React.FC<IconProps>;
-    isOpen?: boolean;
-    cildren?: never;
-    closeIcon?: React.FC<IconProps>;
-    direction?: CollapseIconPropDirection;
-    closeDirection?: CollapseIconPropDirection;
-  },
-  HTMLSpanElement
->;
+import { CollapseIconProps } from '../types';
 
 const cnCollapseIcon = cn('CollapseIcon');
 const animateTimeout = 300;
