+++ src/components/Collapse/__stand__/examples/CollapseExampleLabel/CollapseExampleLabel.css
@@ -1,0 +1,5 @@
+.CollapseExampleLabel {
+  &-Badge {
+    margin-left: var(--space-xs);
+  }
+}
